// backend-server/src/common/enums/review.enum.ts
export enum ReviewType {
  SPACE = 'SPACE',
  PARTNER = 'PARTNER',
  BOOKING = 'BOOKING',
}

export enum ReviewStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  FLAGGED = 'FLAGGED',
}

export enum ReviewRating {
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
}